200% speed increase for toCssString
By using `++` instead of `String.concat`, converting to a css string is
about 200% faster. The reason is list allocation and traversal, which
isn't needed here. I think the code quality does not suffer.

The conversion of colors to strings took about 6% of total time for some svg animations.
It's not a big deal, but an easy gain.